refactor(bench): split profile tests and execution metadata
